pyproject.toml
[project]
name = "fastapi-experment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.30.0",
]
